+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [unreleased] - YYYY-MM-DD
 
+## [0.2.1] - 2019-12-17
+
 ### Added
 
 - Metrics to dead man's switch service.
@@ -32,6 +34,7 @@
 - Alertmanager compatible webhook API.
 - Models and forwarding domain service.
 
-[unreleased]: https://github.com/slok/alertgram/compare/v0.2.0...HEAD
+[unreleased]: https://github.com/slok/alertgram/compare/v0.2.1...HEAD
+[0.2.1]: https://github.com/slok/alertgram/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/slok/alertgram/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/slok/alertgram/releases/tag/v0.1.0